Fill scale constructors for gold & black colors

Fill scale constructors for gold & black colors

scale_color_gb(palette = "main", discrete = TRUE, reverse = FALSE, ...)

Arguments

palette
A character vector indicating the name of palette in gb_palettes. Available palettes:

main: Gold and black colors.

tricol: Gold, black, and dark grey to create a gradual effect.

digitMixed: Five-pack colors specified for digital publications.

printMixed: Five-pack colors specified for printed publications.

full: A palette including all the colors gb_cols can call.

discrete
A logic vector indicating whether color aesthetic is discrete or not; the default is "main".

reverse
A logic vector indicating whether the palette should be reversed.

...
Additional arguments passed to discrete_scale() or scale_color_gradientn(), used respectively when discrete is TRUE or FALSE